django-parser/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-filter>=23.0
django-cors-headers>=4.3
psycopg[binary]>=3.1
